chore: move TreeNode to common data-structures dir
