+++ answers/maximum-depth-of-binary-tree/index.ts
@@ -1,16 +1,4 @@
-/**
- * Definition for a binary tree node.
- */
-export class TreeNode {
-  val: number;
-  left: TreeNode | null;
-  right: TreeNode | null;
-  constructor(val?: number, left?: TreeNode | null, right?: TreeNode | null) {
-    this.val = val === undefined ? 0 : val;
-    this.left = left === undefined ? null : left;
-    this.right = right === undefined ? null : right;
-  }
-}
+import { TreeNode } from '../../data-structures/tree-node';
 
 export function maxDepth(root: TreeNode | null): number {
   if (!root) {
